requirements.txt
langchain
langchain-groq
langchain-community
langchain-core 
langchain-huggingface
langchain-text-splitters
chromadb
python-dotenv
pypdf
rank-bm25
cohere
ragas
datasets
fastapi
uvicorn
gradio
pydantic
sentence-transformers
huggingface-hub

